Added control over CORS proxy for lore
Added a new global setting to control whether the CORS proxy on Glitch is used for loading lore entries. If the user is locally hosting using a server like Simple Web Server that has its own CORS support, the CORS proxy isn't necessary. The side benefit of this is that lore URLs can be hosted on local-network servers that the Glitch CORS proxy can't reach.