Allow multiple replies in each bot query
Add a new menu that allows the user to specify that 1, 2, 3, 4, or 5 replies should be generated each time "send" or "regenerate" is clicked, using the "n" parameter. This saves time and money for users who like to choose between multiple replies, since lore and memories only have to be computed once.

Streaming reply support is turned off when n exceeds 1, because rendering gets really complicated.