Prompt user to reorder messages if needed
If the user has inserted a lot of messages in one place, so that it gets risky to insert new messages in that spot, the user is now prompted to fix the problem using the /respacemessageorders command.

The /respacemessageorders command itself now explains what it's doing, and prompts the user for permission to continue. It then reloads the page afterwards.